Moved fake repositories to the data-testing module
