+++ data-testing/src/main/java/dev/fobo66/core/data/testing/fake/FakeLocationRepository.kt
@@ -1,5 +1,5 @@
 /*
- *    Copyright 2022 Andrey Mukamolov
+ *    Copyright 2025 Andrey Mukamolov
  *
  *    Licensed under the Apache License, Version 2.0 (the "License");
  *    you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package fobo66.valiutchik.domain.fake
+package dev.fobo66.core.data.testing.fake
 
 import fobo66.valiutchik.core.model.repository.LocationRepository
 
@@ -23,6 +23,7 @@ import fobo66.valiutchik.core.model.repository.LocationRepository
  */
 class FakeLocationRepository : LocationRepository {
   var isResolved = false
+
   override suspend fun resolveUserCity(defaultCity: String): String {
     isResolved = true
     return "fake"
